Agregado flujo de caja y financiamiento solucion de problemas en crecimiento anual y correcciones de texto pequeñas

+++ UCR.Negotium/UserControls/CapitalTrabajo.xaml.cs
@@ -64,7 +64,7 @@ namespace UCR.Negotium.UserControls
             }
             set
             {
-                recuperacionCT = Convert.ToInt32(value);
+                recuperacionCT = Convert.ToDouble(value);
             }
         }
